Module in_place::entry

This module allows you to manipulate entries you have acquired

Enums

Entry

An entry which is either Occupied or Vacant.

EntryWithSearchKey

Traits

EntryRemovableOccupiedEntry

A trait to represent recovering to an entry on removal.

InsertableOccupiedEntry

A trait to represent inserting a new entry on top of an existing one.

IntoCollectionMut

KeyedOccupiedEntry

A trait to represent an occupied entry of a collection which owns its keys.

KeyedVacantEntry

A vacant entry which has a key.

NextOccupiedFromOccupied

A trait to extend KeyedOccupiedEntry, allowing the user to get the next occupied entry, according to the ordering on the keys

NextOccupiedFromVacant

A trait to extend KeyedOccupiedEntry, allowing the user to get the next occupied entry, according to the ordering on the keys

OccupiedEntry

A trait to represent an occupied entry of a collection.

OccupiedEntryKeyMut

A trait to extend OccupiedEntry, allowing the user to mutate keys.

PrevOccupiedFromOccupied

A trait to extend KeyedOccupiedEntry, allowing the user to get the previous occupied entry, according to the ordering on the keys

PrevOccupiedFromVacant

A trait to extend KeyedOccupiedEntry, allowing the user to get the previous occupied entry, according to the ordering on the keys

RemovableOccupiedEntry

A trait to represent an OccupiedEntry that can be removed from the collection.

VacantEntry

A vacant entry of a collection.