Trait in_place::entry::VacantEntry

pub trait VacantEntry<'c>: Sized {
    type Value: 'c;
    type Occupied: OccupiedEntry<'c, Value = Self::Value>;
    fn occupy(self, val: Self::Value) -> Self::Occupied;

    fn insert(self, val: Self::Value) { ... }
}

A trait to represent a vacant entry of a collection.

this can be thought of as holding a K, ready to insert when given a value.

The idea is that you’ve done hard work of finding your place in the collection, so inserting shouldn’t be a huge penalty at this point.

Associated Types

type Value: 'c

The type of values in the collection.

type Occupied: OccupiedEntry<'c, Value = Self::Value>

The type of OccupiedEntry we convert to when inserting.

Required methods

fn occupy(self, val: Self::Value) -> Self::Occupied

insert the val using the owned key, returning the occupied entry

Provided methods

fn insert(self, val: Self::Value)

insert the val using the owned key.

if you need to chain a future deletion, use insert_entry